pyproject.toml
[tool.poetry]
name = "elei2022"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.10"
httpx = "^0.23.0"
colorlog = "^6.7.0"
python-dotenv = "^0.21.0"
aiohttp = "^3.8.3"
aiofiles = "^22.1.0"
aiofile = "^3.8.1"
pyunpack = "^0.3"
patool = "^1.12"

[tool.poetry.dev-dependencies]
black = {version = "^22.10.0", allow-prereleases = true}

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
